+++ src/style.css
@@ -10,18 +10,90 @@ body {
     text-align: center;
     width: 100%; 
   }
- /*navegador*/
+
+  /*navegador*/
   .banner
   {  
     width: 100%; 
   }
   nav {
-    background-color: #2395D3;
+    
     color: #fff;
     padding: 10px;
   }
+
+  nav  {
+    list-style: none;
   
- /* Estilos para el formulario de búsqueda */
+    padding: 0;
+    margin: 0;
+  }
+  
+  nav {
+    display: inline-block;
+    margin-right: 10px;
+  }
+  
+  nav a {
+    display: block;
+    
+    text-align: center;
+    padding: 10px;
+    text-decoration: none;
+  }
+
+  nav form {
+    display: inline-block;
+   
+  }
+  
+  nav input[type="text"] {
+    padding: 6px;
+    border: none;
+    border-radius: 4px;
+    margin-top: 8px;
+  }
+  
+  nav button[type="submit"] {
+    padding: 6px 10px;
+    background-color: #0c0c0c;
+    color: #fff;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+  }
+  nav label {
+    display: inline-block;
+    margin-right: 10px;
+  }
+  
+  nav select {
+    padding: 6px;
+    border: none;
+    border-radius: 4px;
+    margin-top: 8px;
+    background-color: #f2f2f2;
+  }
+  
+  nav .btn {
+    display: inline-block;
+   
+  }
+    
+  
+  nav a:hover {
+    background-color: #555;
+  }
+  
+
+
+
+
+
+
+ 
+  
+ /* Estilos para el formulario de búsqueda 
 .searchForm  {
   display: flex;
   align-items: center;
@@ -47,6 +119,10 @@ body {
   background-color: #555;
 }
   
+
+
+
+
 /* Titulo */
   
 h1 {
@@ -132,3 +208,4 @@ footer {
   list-style: none;
   filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
 }
+
